Refactor info box
