+++ mobile/lib/modules/backup/views/backup_controller_page.dart
@@ -6,6 +6,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:immich_mobile/modules/backup/providers/error_backup_list.provider.dart';
+import 'package:immich_mobile/modules/backup/ui/current_backup_asset_info_box.dart';
 import 'package:immich_mobile/modules/login/models/authentication_state.model.dart';
 import 'package:immich_mobile/modules/backup/models/backup_state.model.dart';
 import 'package:immich_mobile/modules/login/providers/authentication.provider.dart';
@@ -467,156 +468,6 @@ class BackupControllerPage extends HookConsumerWidget {
       );
     }
 
-    buildCurrentBackupAssetInfoCard() {
-      String getAssetCreationDate() {
-        return DateFormat.yMMMMd('en_US').format(
-          DateTime.parse(
-            backupState.currentUploadAsset.createdAt.toString(),
-          ).toLocal(),
-        );
-      }
-
-      return ListTile(
-        leading: Icon(
-          Icons.info_outline_rounded,
-          color: Theme.of(context).primaryColor,
-        ),
-        title: Row(
-          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-          children: [
-            const Text(
-              "backup_controller_page_uploading_file_info",
-              style: TextStyle(fontWeight: FontWeight.bold, fontSize: 14),
-            ).tr(),
-            if (ref.watch(errorBackupListProvider).isNotEmpty)
-              ActionChip(
-                avatar: Icon(
-                  Icons.info,
-                  color: Colors.red[400],
-                ),
-                elevation: 1,
-                visualDensity: VisualDensity.compact,
-                label: Text(
-                  "backup_controller_page_failed",
-                  style: TextStyle(
-                    color: Colors.red[400],
-                    fontWeight: FontWeight.bold,
-                    fontSize: 11,
-                  ),
-                ).tr(
-                  args: [ref.watch(errorBackupListProvider).length.toString()],
-                ),
-                backgroundColor: Colors.white,
-                onPressed: () {
-                  AutoRouter.of(context).push(const FailedBackupStatusRoute());
-                },
-              ),
-          ],
-        ),
-        subtitle: Column(
-          children: [
-            Padding(
-              padding: const EdgeInsets.only(top: 8.0),
-              child: Row(
-                children: [
-                  Expanded(
-                    child: LinearProgressIndicator(
-                      minHeight: 10.0,
-                      value: backupState.progressInPercentage / 100.0,
-                      backgroundColor: Colors.grey,
-                      color: Theme.of(context).primaryColor,
-                    ),
-                  ),
-                  Text(
-                    " ${backupState.progressInPercentage.toStringAsFixed(0)}%",
-                    style: const TextStyle(fontSize: 12),
-                  )
-                ],
-              ),
-            ),
-            Padding(
-              padding: const EdgeInsets.only(top: 8.0),
-              child: Table(
-                border: TableBorder.all(
-                  color: Theme.of(context).primaryColorLight,
-                  width: 1,
-                ),
-                children: [
-                  TableRow(
-                    decoration: const BoxDecoration(
-                        // color: Colors.grey[100],
-                        ),
-                    children: [
-                      TableCell(
-                        verticalAlignment: TableCellVerticalAlignment.middle,
-                        child: Padding(
-                          padding: const EdgeInsets.all(6.0),
-                          child: const Text(
-                            'backup_controller_page_filename',
-                            style: TextStyle(
-                              fontWeight: FontWeight.bold,
-                              fontSize: 10.0,
-                            ),
-                          ).tr(
-                            args: [
-                              backupState.currentUploadAsset.fileName,
-                              backupState.currentUploadAsset.fileType
-                                  .toLowerCase()
-                            ],
-                          ),
-                        ),
-                      ),
-                    ],
-                  ),
-                  TableRow(
-                    decoration: const BoxDecoration(
-                        // color: Colors.grey[200],
-                        ),
-                    children: [
-                      TableCell(
-                        verticalAlignment: TableCellVerticalAlignment.middle,
-                        child: Padding(
-                          padding: const EdgeInsets.all(6.0),
-                          child: const Text(
-                            "backup_controller_page_created",
-                            style: TextStyle(
-                              fontWeight: FontWeight.bold,
-                              fontSize: 10.0,
-                            ),
-                          ).tr(
-                            args: [getAssetCreationDate()],
-                          ),
-                        ),
-                      ),
-                    ],
-                  ),
-                  TableRow(
-                    decoration: const BoxDecoration(
-                        // color: Colors.grey[100],
-                        ),
-                    children: [
-                      TableCell(
-                        child: Padding(
-                          padding: const EdgeInsets.all(6.0),
-                          child: const Text(
-                            "backup_controller_page_id",
-                            style: TextStyle(
-                              fontWeight: FontWeight.bold,
-                              fontSize: 10.0,
-                            ),
-                          ).tr(args: [backupState.currentUploadAsset.id]),
-                        ),
-                      ),
-                    ],
-                  ),
-                ],
-              ),
-            ),
-          ],
-        ),
-      );
-    }
-
     void startBackup() {
       ref.watch(errorBackupListProvider.notifier).empty();
       if (ref.watch(backupProvider).backupProgress !=
@@ -700,7 +551,7 @@ class BackupControllerPage extends HookConsumerWidget {
             const Divider(),
             buildStorageInformation(),
             const Divider(),
-            buildCurrentBackupAssetInfoCard(),
+            const CurrentUploadingAssetInfoBox(),
             Padding(
               padding: const EdgeInsets.only(
                 top: 24,
